get MongoDatabase in working order

+++ WebApi/Entities/Location.cs
@@ -7,7 +7,7 @@ namespace Web_Api.Entities
 {
     public class Location
     {
-        public float lat;
-        public float lon;
+        public double lat;
+        public double lon;
     }
 }
